Added alert call when existing username is added again
While appropriate checks were included to prevent users from entering
and saving an existing username, there was no visual feedback for the
user to indicate this. And so, code has been added to an existing
if check to alert users if they add a username that already exists in
the Set. A simple alert call was used.