Added alert for when user types in invalid username
Entering an invalid username will cause a log to the error console
without any visual feedback to the user, so added is a check within
the second then method of the fetch call to determine the status of the
fetch response and the details if it fails. If the response status is
"Failed" and the type of failure is "Invalid username", the user will
be alerted using the alert call.