Add username in correct format to card's dataset during creation
The username from the API fetch response comes in a inconsistent format.
When the user is ranked, only the username is returned; when the user is
unranked the username comes in a different format (username:_username_).

So to make the format consistent, in the `createCard` function, the
returned username is first split. Then the last element of the resulting
array is used, making sure only the actualy username is grabbed. The
resulting username is then saved into the card's dataset so it's easy
to grab later on, especially for deleting.