Fix deleteUser bug so user actually gets deleted from local storage
Previously, in the `deleteUser` function, the value used to delete
a uesr from the set is extracted from the card's `.username` element's
innerHTML, which is in an incorrect format, making the delete operation
not do anything. Subsequently, the actual item from the local storage
isn't deleted--only the element is.

To fix this, the value used to delete user from the set is gotten from
the card's dataset, which contains the correct username.