Fix useEffect dependency warning
Problem: useEffect for when the App compnent mounts is giving a
missing dependency warning because it uses the fetchData function
defined inside the component.

The solution is to wrap the necessary dependencies inside a useCallback
and declaring them inside the dependency array(s).This way, the
useEffect when the app mounts will still have the desired
behaviour of running only once, this time without the warning.

The first change was to wrap fetchData inside a useCallback, which
depends on the removeUser function. So removeUser function is also
wrapped inside a useCallback. removeUser contains a state dependency
with `users`, which would have caused an infinite loop. So the dependency
was removed. Instead, I used a function inside of the `setUsers` method
to update the state and the localStorage.